+++ tritium/src/main/java/net/minecraft/block/BlockCauldron.java
@@ -72,7 +72,7 @@ public class BlockCauldron extends Block {
      * Called When an Entity Collided with the Block
      */
     public void onEntityCollidedWithBlock(World worldIn, BlockPos pos, IBlockState state, Entity entityIn) {
-        int i = state.getValue(LEVEL).intValue();
+        int i = state.getValue(LEVEL);
         float f = (float) pos.getY() + (6.0F + (float) (3 * i)) / 16.0F;
 
         if (!worldIn.isRemote && entityIn.isBurning() && i > 0 && entityIn.getEntityBoundingBox().minY <= (double) f) {
@@ -90,7 +90,7 @@ public class BlockCauldron extends Block {
             if (itemstack == null) {
                 return true;
             } else {
-                int i = state.getValue(LEVEL).intValue();
+                int i = state.getValue(LEVEL);
                 Item item = itemstack.getItem();
 
                 if (item == Items.water_bucket) {
@@ -181,7 +181,7 @@ public class BlockCauldron extends Block {
         if (worldIn.rand.nextInt(20) == 1) {
             IBlockState iblockstate = worldIn.getBlockState(pos);
 
-            if (iblockstate.getValue(LEVEL).intValue() < 3) {
+            if (iblockstate.getValue(LEVEL) < 3) {
                 worldIn.setBlockState(pos, iblockstate.cycleProperty(LEVEL), 2);
             }
         }
@@ -203,7 +203,7 @@ public class BlockCauldron extends Block {
     }
 
     public int getComparatorInputOverride(World worldIn, BlockPos pos) {
-        return worldIn.getBlockState(pos).getValue(LEVEL).intValue();
+        return worldIn.getBlockState(pos).getValue(LEVEL);
     }
 
     /**
@@ -217,7 +217,7 @@ public class BlockCauldron extends Block {
      * Convert the BlockState into the correct metadata value
      */
     public int getMetaFromState(IBlockState state) {
-        return state.getValue(LEVEL).intValue();
+        return state.getValue(LEVEL);
     }
 
     protected BlockState createBlockState() {
