+++ tritium/src/main/java/net/minecraft/block/BlockCrops.java
@@ -38,7 +38,7 @@ public class BlockCrops extends BlockBush implements IGrowable {
         super.updateTick(worldIn, pos, state, rand);
 
         if (worldIn.getLightFromNeighbors(pos.up()) >= 9) {
-            int i = state.getValue(AGE).intValue();
+            int i = state.getValue(AGE);
 
             if (i < 7) {
                 float f = getGrowthChance(this, worldIn, pos);
@@ -51,7 +51,7 @@ public class BlockCrops extends BlockBush implements IGrowable {
     }
 
     public void grow(World worldIn, BlockPos pos, IBlockState state) {
-        int i = state.getValue(AGE).intValue() + MathHelper.getRandomIntegerInRange(worldIn.rand, 2, 5);
+        int i = state.getValue(AGE) + MathHelper.getRandomIntegerInRange(worldIn.rand, 2, 5);
 
         if (i > 7) {
             i = 7;
@@ -72,7 +72,7 @@ public class BlockCrops extends BlockBush implements IGrowable {
                 if (iblockstate.getBlock() == Blocks.farmland) {
                     f1 = 1.0F;
 
-                    if (iblockstate.getValue(BlockFarmland.MOISTURE).intValue() > 0) {
+                    if (iblockstate.getValue(BlockFarmland.MOISTURE) > 0) {
                         f1 = 3.0F;
                     }
                 }
@@ -124,7 +124,7 @@ public class BlockCrops extends BlockBush implements IGrowable {
         super.dropBlockAsItemWithChance(worldIn, pos, state, chance, 0);
 
         if (!worldIn.isRemote) {
-            int i = state.getValue(AGE).intValue();
+            int i = state.getValue(AGE);
 
             if (i >= 7) {
                 int j = 3 + fortune;
@@ -142,7 +142,7 @@ public class BlockCrops extends BlockBush implements IGrowable {
      * Get the Item that this Block should drop when harvested.
      */
     public Item getItemDropped(IBlockState state, Random rand, int fortune) {
-        return state.getValue(AGE).intValue() == 7 ? this.getCrop() : this.getSeed();
+        return state.getValue(AGE) == 7 ? this.getCrop() : this.getSeed();
     }
 
     public Item getItem(World worldIn, BlockPos pos) {
@@ -153,7 +153,7 @@ public class BlockCrops extends BlockBush implements IGrowable {
      * Whether this IGrowable can grow
      */
     public boolean canGrow(World worldIn, BlockPos pos, IBlockState state, boolean isClient) {
-        return state.getValue(AGE).intValue() < 7;
+        return state.getValue(AGE) < 7;
     }
 
     public boolean canUseBonemeal(World worldIn, Random rand, BlockPos pos, IBlockState state) {
@@ -175,7 +175,7 @@ public class BlockCrops extends BlockBush implements IGrowable {
      * Convert the BlockState into the correct metadata value
      */
     public int getMetaFromState(IBlockState state) {
-        return state.getValue(AGE).intValue();
+        return state.getValue(AGE);
     }
 
     protected BlockState createBlockState() {
