+++ tritium/src/main/java/net/minecraft/block/BlockReed.java
@@ -36,7 +36,7 @@ public class BlockReed extends Block {
                 }
 
                 if (i < 3) {
-                    int j = state.getValue(AGE).intValue();
+                    int j = state.getValue(AGE);
 
                     if (j == 15) {
                         worldIn.setBlockState(pos.up(), this.getDefaultState());
@@ -133,7 +133,7 @@ public class BlockReed extends Block {
      * Convert the BlockState into the correct metadata value
      */
     public int getMetaFromState(IBlockState state) {
-        return state.getValue(AGE).intValue();
+        return state.getValue(AGE);
     }
 
     protected BlockState createBlockState() {
