+++ tritium/src/main/java/net/minecraft/block/BlockSnow.java
@@ -33,11 +33,11 @@ public class BlockSnow extends Block {
     }
 
     public boolean isPassable(IBlockAccess worldIn, BlockPos pos) {
-        return worldIn.getBlockState(pos).getValue(LAYERS).intValue() < 5;
+        return worldIn.getBlockState(pos).getValue(LAYERS) < 5;
     }
 
     public AxisAlignedBB getCollisionBoundingBox(World worldIn, BlockPos pos, IBlockState state) {
-        int i = state.getValue(LAYERS).intValue() - 1;
+        int i = state.getValue(LAYERS) - 1;
         float f = 0.125F;
         return new AxisAlignedBB((double) pos.getX() + this.minX, (double) pos.getY() + this.minY, (double) pos.getZ() + this.minZ, (double) pos.getX() + this.maxX, (float) pos.getY() + (float) i * f, (double) pos.getZ() + this.maxZ);
     }
@@ -62,7 +62,7 @@ public class BlockSnow extends Block {
 
     public void setBlockBoundsBasedOnState(IBlockAccess worldIn, BlockPos pos) {
         IBlockState iblockstate = worldIn.getBlockState(pos);
-        this.getBoundsForLayers(iblockstate.getValue(LAYERS).intValue());
+        this.getBoundsForLayers(iblockstate.getValue(LAYERS));
     }
 
     protected void getBoundsForLayers(int p_150154_1_) {
@@ -72,7 +72,7 @@ public class BlockSnow extends Block {
     public boolean canPlaceBlockAt(World worldIn, BlockPos pos) {
         IBlockState iblockstate = worldIn.getBlockState(pos.down());
         Block block = iblockstate.getBlock();
-        return block != Blocks.ice && block != Blocks.packed_ice && (block.getMaterial() == Material.leaves || (block == this && iblockstate.getValue(LAYERS).intValue() >= 7 || block.isOpaqueCube() && block.blockMaterial.blocksMovement()));
+        return block != Blocks.ice && block != Blocks.packed_ice && (block.getMaterial() == Material.leaves || (block == this && iblockstate.getValue(LAYERS) >= 7 || block.isOpaqueCube() && block.blockMaterial.blocksMovement()));
     }
 
     /**
@@ -93,7 +93,7 @@ public class BlockSnow extends Block {
     }
 
     public void harvestBlock(World worldIn, EntityPlayer player, BlockPos pos, IBlockState state, TileEntity te) {
-        spawnAsEntity(worldIn, pos, new ItemStack(Items.snowball, state.getValue(LAYERS).intValue() + 1, 0));
+        spawnAsEntity(worldIn, pos, new ItemStack(Items.snowball, state.getValue(LAYERS) + 1, 0));
         worldIn.setBlockToAir(pos);
     }
 
@@ -133,14 +133,14 @@ public class BlockSnow extends Block {
      * Whether this Block can be replaced directly by other blocks (true for e.g. tall grass)
      */
     public boolean isReplaceable(World worldIn, BlockPos pos) {
-        return worldIn.getBlockState(pos).getValue(LAYERS).intValue() == 1;
+        return worldIn.getBlockState(pos).getValue(LAYERS) == 1;
     }
 
     /**
      * Convert the BlockState into the correct metadata value
      */
     public int getMetaFromState(IBlockState state) {
-        return state.getValue(LAYERS).intValue() - 1;
+        return state.getValue(LAYERS) - 1;
     }
 
     protected BlockState createBlockState() {
