+++ tritium/src/main/java/net/minecraft/command/CommandReplaceItem.java
@@ -149,7 +149,7 @@ public class CommandReplaceItem extends CommandBase {
         if (!SHORTCUTS.containsKey(shortcut)) {
             throw new CommandException("commands.generic.parameter.invalid", shortcut);
         } else {
-            return SHORTCUTS.get(shortcut).intValue();
+            return SHORTCUTS.get(shortcut);
         }
     }
