+++ tritium/src/main/java/net/minecraft/enchantment/EnchantmentHelper.java
@@ -84,7 +84,7 @@ public class EnchantmentHelper {
         Iterator iterator = enchMap.keySet().iterator();
 
         while (iterator.hasNext()) {
-            int i = ((Integer) iterator.next()).intValue();
+            int i = (Integer) iterator.next();
             Enchantment enchantment = Enchantment.getEnchantmentById(i);
 
             if (enchantment != null) {
@@ -94,7 +94,7 @@ public class EnchantmentHelper {
                 nbttaglist.appendTag(nbttagcompound);
 
                 if (stack.getItem() == Items.enchanted_book) {
-                    Items.enchanted_book.addEnchantment(stack, new EnchantmentData(enchantment, enchMap.get(i).intValue()));
+                    Items.enchanted_book.addEnchantment(stack, new EnchantmentData(enchantment, enchMap.get(i)));
                 }
             }
         }
@@ -385,7 +385,7 @@ public class EnchantmentHelper {
                             boolean flag = true;
 
                             for (EnchantmentData enchantmentdata1 : list) {
-                                if (!enchantmentdata1.enchantmentobj.canApplyTogether(Enchantment.getEnchantmentById(integer.intValue()))) {
+                                if (!enchantmentdata1.enchantmentobj.canApplyTogether(Enchantment.getEnchantmentById(integer))) {
                                     flag = false;
                                     break;
                                 }
