+++ tritium/src/main/java/net/minecraft/entity/EntityList.java
@@ -135,7 +135,7 @@ public class EntityList {
      */
     public static int getEntityID(Entity entityIn) {
         Integer integer = classToIDMapping.get(entityIn.getClass());
-        return integer == null ? 0 : integer.intValue();
+        return integer == null ? 0 : integer;
     }
 
     public static Class<? extends Entity> getClassFromID(int entityID) {
@@ -154,7 +154,7 @@ public class EntityList {
      */
     public static int getIDFromString(String entityName) {
         Integer integer = stringToIDMapping.get(entityName);
-        return integer == null ? 90 : integer.intValue();
+        return integer == null ? 90 : integer;
     }
 
     /**
